pyproject.toml
[project]
name = "babyclaw-server"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.122.0",
  "uvicorn[standard]>=0.38.0",
  "boto3>=1.41.0",
  "python-dotenv>=1.0.0",
  "pydantic>=2.12.5",
  "strands-agents>=1.18.0",
  "strands-agents-tools>=0.2.0"
]
